+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## dbt-databricks 1.11.0 (TBD)
+
+### Features
+
+- Support databricks_tags for MV/STs
+
 ## dbt-databricks 1.10.11 (TBD)
 
 ### Fixes